add common dir
