+++ andriuschin.danil/common/format_guard.cpp
