+++ andriuschin.danil/common/format_guard.hpp
